add create-change-detail.js, update
