+++ utils/index.js
@@ -1,5 +1,6 @@
 
 module.exports = {
     'ChangeDetail': require('./change-detail.js'),
+    'createChangeDetail': require('./create-change-detail.js'),
     '__proto__': null
 }
